Move docker arguments to internal modules
The modules for handling docker arguments have been moved to the
directory for internal modules, as they should not be exposed to users
of the library.
