--- dampf/lib/Dampf/Docker/Args.hs
@@ -1,10 +1,0 @@
-module Dampf.Docker.Args
-  ( ToArgs(..)
-  , RunArgs
-  , HasRunArgs(..)
-  , mkRunArgs
-  ) where
-
-import           Dampf.Docker.Args.Class
-import           Dampf.Docker.Args.Run
-
